--- errors/InvalidFieldsError.js
@@ -1,8 +1,0 @@
-'use strict'
-
-const BaseError = require('./BaseError')
-
-class InvalidFieldsError extends BaseError {
-}
-
-module.exports = InvalidFieldsError
