--- errors/TransferHandlerAlreadyRegisteredError.js
@@ -1,8 +1,0 @@
-'use strict'
-
-const BaseError = require('./BaseError')
-
-class TransferHandlerAlreadyRegisteredError extends BaseError {
-}
-
-module.exports = TransferHandlerAlreadyRegisteredError
