+++ src/errors/InvalidFieldsError.ts
@@ -1,0 +1,4 @@
+import { BaseError } from './BaseError'
+
+export class InvalidFieldsError extends BaseError {
+}
