+++ src/errors/TransferHandlerAlreadyRegisteredError.ts
@@ -1,0 +1,4 @@
+import { BaseError } from './BaseError'
+
+export class TransferHandlerAlreadyRegisteredError extends BaseError {
+}
