+++ src/errors/index.ts
@@ -1,0 +1,4 @@
+export { BaseError } from './BaseError'
+export { InterledgerRejectionError } from './InterledgerRejectionError'
+export { InvalidFieldsError } from './InvalidFieldsError'
+export { TransferHandlerAlreadyRegisteredError } from './TransferHandlerAlreadyRegisteredError'
